Task_branch_48: тестовая ветка(реализовано отображение списка категорий)
