+++ app/src/main/java/com/example/pocketstorage/presentation/ui/screens/category/CategoriesStateForCurrentLocation.kt
@@ -1,0 +1,9 @@
+package com.example.pocketstorage.presentation.ui.screens.category
+
+import com.example.pocketstorage.domain.model.Category
+
+data class CategoriesStateForCurrentLocation(
+    val currentLocationId: String = "",
+    val existingCategoriesForCurrentLocation: List<Category> = emptyList(),
+    val searchText: String = ""
+)
